Move modules from `stellargraph.utils` to top-level ones (#938)
This moves most of the code within `stellargraph.utils` to dedicated top-level
modules, because `utils` is an overly generic name.

The old paths still work, but have runtime warnings to aid the migration. This
is done by overriding `__getattr__` for the `stellargraph.utils` module
(unfortunately this cannot use [PEP
562](https://www.python.org/dev/peps/pep-0562/) because that is Python 3.7+, and
we support 3.6), and replacing each of the submodule files with, essentially:

```python
warnings.warn(...)
from path.to.new_module import *
```

(Unfortunately, I think these have to be files with the same name, so these new
small files are stopping git from detecting the moves: all of the large new
files are exact moves, with no other changes.)

These two forms are both needed because there's two possibilities for how users
might be using this:

1. `import stellargraph.utils` followed by `utils.foo`: calls `__getattr__` if
   it's not otherwise found
2. a direct import of a submodule `import stellargraph.utils.foo`: I believe
   this just does a file system traversal and opens/executes the given file
   (i.e. `stellargraph/utils/foo.py` or `stellargraph/utils/foo/__init__.py`) to
   create the submodule directly (rather than getting attributes of parent
   modules), and so needs the handling in each file individually


Full list of changes:

| old                                           | new                                                                  |
|-----------------------------------------------|----------------------------------------------------------------------|
| stellargraph.utils.calibration                | stellargraph.calibration                                             |
| stellargraph.utils.IsotonicCalibration        | stellargraph.calibration.IsotonicCalibration                         |
| stellargraph.utils.TemperatureCalibration     | stellargraph.calibration.TemperatureCalibration                      |
| stellargraph.utils.expected_calibration_error | stellargraph.calibration.expected_calibration_error                  |
| stellargraph.utils.plot_reliability_diagram   | stellargraph.calibration.plot_reliability_diagram                    |
| stellargraph.utils.ensemble                   | stellargraph.ensemble                                                |
| stellargraph.utils.BaggingEnsemble            | stellargraph.ensemble.BaggingEnsemble                                |
| stellargraph.utils.Ensemble                   | stellargraph.ensemble.Ensemble                                       |
| stellargraph.utils.saliency_maps              | stellargraph.interpretability.saliency_maps                          |
| stellargraph.utils.integrated_gradients       | stellargraph.interpretability.saliency_maps.integrated_gradients     |
| stellargraph.utils.integrated_gradients_gat   | stellargraph.interpretability.saliency_maps.integrated_gradients_gat |
| stellargraph.utils.saliency_gat               | stellargraph.interpretability.saliency_maps.saliency_gat             |
| stellargraph.utils.GradientSaliencyGAT        | stellargraph.interpretability.GradientSaliencyGAT                    |
| stellargraph.utils.IntegratedGradients        | stellargraph.interpretability.IntegratedGradients                    |
| stellargraph.utils.IntegratedGradientsGAT     | stellargraph.interpretability.IntegratedGradientsGAT                 |

Computed by adding the following to `stellargraph/utils/__init__.py`:

```
for name, (new_module_name, new_value) in sorted(
    _MAPPING.items(),
    key=lambda x: ("stellargraph." + x[1][0] + "zzz", x[0]) if x[1][0] is not None else (x[1][1].__name__, "")
):
    if isinstance(new_value, types.ModuleType):
        new_location = new_value.__name__
    else:
        new_location = f"stellargraph.{new_module_name}.{new_value.__name__}"

    print(f"| `stellargraph.utils.{name}` | `{new_location}` |")

raise ValueError()
```

See: #909
