+++ docs/api.txt
@@ -109,14 +109,14 @@ Link prediction layers
 Ensembles
 ------------------------
 
-.. automodule:: stellargraph.utils.ensemble
+.. automodule:: stellargraph.ensemble
   :members: Ensemble, BaggingEnsemble
 
 
 Calibration
 ------------------------
 
-.. automodule:: stellargraph.utils.calibration
+.. automodule:: stellargraph.calibration
   :members: expected_calibration_error, plot_reliability_diagram, TemperatureCalibration, IsotonicCalibration, fit, predict
 
 Utilities
