+++ stellargraph/interpretability/saliency_maps/__init__.py
@@ -14,11 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import warnings as warnings
-
-warnings.warn(
-    "'stellargraph.utils.saliency_maps.integrated_gradients' has been moved to 'stellargraph.interpretability.saliency_maps.integrated_gradients'",
-    DeprecationWarning,
-)
-
-from ...interpretability.saliency_maps.integrated_gradients import *
+from .integrated_gradients import IntegratedGradients
+from .integrated_gradients_gat import IntegratedGradientsGAT
+from .saliency_gat import GradientSaliencyGAT
