+++ tests/interpretability/__init__.py
@@ -13,12 +13,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-import warnings as warnings
-
-warnings.warn(
-    "'stellargraph.utils.saliency_maps.integrated_gradients' has been moved to 'stellargraph.interpretability.saliency_maps.integrated_gradients'",
-    DeprecationWarning,
-)
-
-from ...interpretability.saliency_maps.integrated_gradients import *
